requirements.txt
wikipedia
transformers
sentence-transformers
faiss-cpu  # or faiss-gpu if you have CUDA
numpy